feat(Pymake): add option to build from inplace source files
Additional cleanup of unused methods and consistent methods to strip
paths and extensions from targets and compilers.
