+++ examples/make_mfusg.py
@@ -2,7 +2,7 @@ import pymake
 
 
 def make_mfusg():
-    pymake.build_apps(["mfusg", "zonbudusg"])
+    pymake.build_apps(["mfusg"])
 
 
 if __name__ == "__main__":
